Split single::ent mod
